backend/requirements.txt
fastapi
sqlalchemy
databases
pydantic
uvicorn
passlib[bcrypt]
python-dotenv
aiomysql
pyjwt
pytest
pytest-asyncio
pytest-mock
httpx